feat (msg reply): make chatbox resize on input
This adds an IconTextArea to the ui_kit and uses it for the modal
window in compose/msg/mod.rs, which appears when replying to a
message.

The `text` field is lifted out of the TextArea and IconTextArea
components so that a submit button can clear them. I haven't
figured out why the submit button on the "reply" modal doesn't
work, so that functionality has been left unchanged.

TESTING: to test this change, do the following:
1) send or receive a message
2) click on that message
3) in the modal which appears, select the text box and hold down the
'a' key. as the number of characters fill the box, the height of the
box should increase. as characters are deleted, the height should
decrease. pressing the 'enter' key should clear the box.