fix: Prevent rerendering by setting state once and updating accordingly. (#102)
* fix: Prevent rerendering by setting state once and updating accordingly.

Although additional changes will be needed in other components, this is the step towards preventing rerendering by updating the state only when it needs to be updated. Some parts will use Scope::spawn to push the update into a separate task.

* chore: Set states for friends

Note: `Popup` component would still cause rerendering, and thus would cause `Friends` to be rerendered as well

* chore: Minor cleanup

* chore: Remove unused import

* chore: Move friends count to a state

* chore: Move reqCount to a state and update it in a separate task

* fix: Use use_future instead of spawn

Base on the code behind this function, it should cancel the task running in the scope instead of spawning more task on each render.

* chore: Implement DerefMut for Account and Messaging holders

* refactor: Change the messages component (WIP)

* chore: Cleanup codebase

* refactor: Move messages to be apart of `compose/mod.rs` and send messages directly to the component

Note: This broke conversation switching for the tiem being

* chore: Move having messages back into its own component

* chore: Replace unnecessary call with `messages`

* chore: Remove unused variables and imports

* chore: Remove unused import

* fix: Use milliseconds

* chore (formatting): run cargo fmt

* add error logging for debug purposes

* fix (messaging): properly share state between components

This change removes the cloned `Conversation` struct from the props
passed to Compose, Topbar, and messages. They now call `use_atom_ref`
to get the global state, and are now re-rendered when a conversation
is selected

* fix: comment out debug code

* chore (clippy): make changes that clippy would suggest

Co-authored-by: Stuart Woodbury <stuartwoodbury@protonmail.ch>