fix (message replies): shrink height on send button or close modal
this change re-introduces some javascript that shouldn't have been
deleted. it also closes the msg reply modal when the reply button
is pressed, because it was too hard to make a JS function to handle
both the send and reply buttons from within the TextArea.

TESTING: JS isn't being used to interact with the input. hopefully this
fixes the issue of double sending messages when spamming the button.
See if this is true.